Move implementations to its own file, so there is more clarity
